+++ src/main/kotlin/org/ionproject/integration/application/job/ISELAcademicCalendarJob.kt
@@ -4,14 +4,14 @@ import org.ionproject.integration.application.config.AppProperties
 import org.ionproject.integration.application.job.tasklet.DownloadAndCompareTasklet
 import org.ionproject.integration.application.dto.AcademicCalendarData
 import org.ionproject.integration.application.dispatcher.IDispatcher
-import org.ionproject.integration.infrastructure.OutputFormat
+import org.ionproject.integration.infrastructure.file.OutputFormat
 import org.ionproject.integration.infrastructure.pdfextractor.AcademicCalendarExtractor
 import org.ionproject.integration.infrastructure.pdfextractor.ITextPdfExtractor
-import org.ionproject.integration.infrastructure.FileComparatorImpl
-import org.ionproject.integration.infrastructure.FileDigestImpl
-import org.ionproject.integration.infrastructure.PDFBytesFormatChecker
-import org.ionproject.integration.infrastructure.IFileDownloader
-import org.ionproject.integration.infrastructure.HashRepositoryImpl
+import org.ionproject.integration.infrastructure.file.FileComparatorImpl
+import org.ionproject.integration.infrastructure.file.FileDigestImpl
+import org.ionproject.integration.infrastructure.pdfextractor.PDFBytesFormatChecker
+import org.ionproject.integration.infrastructure.http.IFileDownloader
+import org.ionproject.integration.infrastructure.hash.HashRepositoryImpl
 import org.ionproject.integration.model.external.calendar.AcademicCalendar
 import org.ionproject.integration.model.external.calendar.AcademicCalendarDto
 import org.ionproject.integration.domain.calendar.RawCalendarData
