+++ src/main/kotlin/org/ionproject/integration/infrastructure/repository/IntegrationJobRepository.kt
@@ -1,7 +1,7 @@
 package org.ionproject.integration.infrastructure.repository
 
 import org.ionproject.integration.application.JobEngine
-import org.ionproject.integration.infrastructure.OutputFormat
+import org.ionproject.integration.infrastructure.file.OutputFormat
 import org.ionproject.integration.application.job.JobType
 import org.springframework.jdbc.core.JdbcTemplate
 import org.springframework.jdbc.core.ResultSetExtractor
