+++ src/main/kotlin/org/ionproject/integration/ui/dto/SafeTimetableJobDto.kt
@@ -1,7 +1,7 @@
 package org.ionproject.integration.ui.dto
 
 import org.ionproject.integration.application.JobEngine
-import org.ionproject.integration.infrastructure.OutputFormat
+import org.ionproject.integration.infrastructure.file.OutputFormat
 import org.ionproject.integration.infrastructure.repository.IInstitutionRepository
 import org.ionproject.integration.infrastructure.repository.IProgrammeRepository
