+++ src/test/kotlin/org/ionproject/integration/dispatcher/DispatcherTests.kt
@@ -4,10 +4,11 @@ import org.ionproject.integration.application.config.AppProperties
 import org.ionproject.integration.application.dispatcher.DispatcherImpl
 import org.ionproject.integration.application.dispatcher.DispatchResult
 import org.ionproject.integration.application.dto.ParsedData
-import org.ionproject.integration.infrastructure.GitOutcome
-import org.ionproject.integration.infrastructure.IGitHandler
-import org.ionproject.integration.infrastructure.IGitHandlerFactory
-import org.ionproject.integration.infrastructure.OutputFormat
+import org.ionproject.integration.infrastructure.git.GitOutcome
+import org.ionproject.integration.infrastructure.git.IGitHandler
+import org.ionproject.integration.infrastructure.git.IGitHandlerFactory
+import org.ionproject.integration.infrastructure.file.OutputFormat
+import org.ionproject.integration.infrastructure.file.IFileWriter
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
 import org.mockito.kotlin.any
