+++ src/test/kotlin/org/ionproject/integration/file/implementations/BytesFormatCheckerTest.kt
@@ -1,6 +1,6 @@
 package org.ionproject.integration.file.implementations
 
-import org.ionproject.integration.infrastructure.PDFBytesFormatChecker
+import org.ionproject.integration.infrastructure.pdfextractor.PDFBytesFormatChecker
 import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
