+++ src/test/kotlin/org/ionproject/integration/format/implementations/StringFormatCheckerTest.kt
@@ -1,6 +1,6 @@
 package org.ionproject.integration.format.implementations
 
-import org.ionproject.integration.infrastructure.StringFormatChecker
+import org.ionproject.integration.infrastructure.text.StringFormatChecker
 import org.junit.jupiter.api.Assertions.assertFalse
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
