+++ src/test/kotlin/org/ionproject/integration/utils/RegexUtilsTests.kt
@@ -1,6 +1,6 @@
 package org.ionproject.integration.utils
 
-import org.ionproject.integration.infrastructure.RegexUtils
+import org.ionproject.integration.infrastructure.text.RegexUtils
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
